Update 'update' method in ProyectoServiceImpl
now when updating, the student status of the project can be changed from
active to inactive or vice versa. To achieve this, the array 'estudiantes'
in ProyectoCreationDTO it must have the ID of the student that the user wants to set as active,
i.e. if project A has students ab12345 and cd06789  and these are active,
if the user wants to update student ab12345 to inactive, the array 'estudiantes'
must only have student cd06789 as member. When the service method
captures only this id, it will set inactive the absent student